+++ src/web/driver/webgl/ScanlineProcessor.ts
@@ -6,6 +6,8 @@ class ScanlineProcessor implements Processor {
     constructor(private _gl: WebGLRenderingContext) {}
 
     init(): void {
+        if (this._initialized) return;
+
         const gl = this._gl;
 
         this._framebuffer = gl.createFramebuffer();
@@ -22,9 +24,13 @@ class ScanlineProcessor implements Processor {
 
         gl.bindBuffer(gl.ARRAY_BUFFER, this._textureCoordinateBuffer);
         gl.bufferData(gl.ARRAY_BUFFER, new Float32Array([1, 1, 0, 1, 1, 0, 0, 0]), gl.STATIC_DRAW);
+
+        this._initialized = true;
     }
 
     destroy(): void {
+        if (!this._initialized) return;
+
         const gl = this._gl;
 
         this._program.delete();
@@ -33,6 +39,8 @@ class ScanlineProcessor implements Processor {
         gl.deleteBuffer(this._textureCoordinateBuffer);
 
         if (this._texture) gl.deleteTexture(this._texture);
+
+        this._initialized = false;
     }
 
     render(texture: WebGLTexture): void {
@@ -87,10 +95,11 @@ class ScanlineProcessor implements Processor {
         return this._texture;
     }
 
-    configure(width: number, height: number, level: number): void {
+    resize(width: number, height: number): void {
+        if (!this._initialized) return;
+
         this._width = width;
         this._height = height;
-        this._level = level;
 
         const gl = this._gl;
 
@@ -108,19 +117,26 @@ class ScanlineProcessor implements Processor {
         gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
 
         this._program.use();
-        this._program.uniform1f(fsh.scanlines.uniform.level, 1 - this._level);
         this._program.uniform1f(fsh.scanlines.uniform.height, 2 * this._height);
+    }
+
+    configure(level: number): void {
+        if (!this._initialized) return;
+
+        this._program.use();
+        this._program.uniform1f(fsh.scanlines.uniform.level, 1 - level);
     }
 
     private _width = 0;
     private _height = 0;
-    private _level = 0;
 
     private _texture: WebGLTexture = null;
     private _program: Program = null;
     private _framebuffer: WebGLFramebuffer = null;
     private _vertexCoordinateBuffer: WebGLBuffer = null;
     private _textureCoordinateBuffer: WebGLBuffer = null;
+
+    private _initialized = false;
 }
 
 export default ScanlineProcessor;
